+++ conf.lua
@@ -1,3 +1,6 @@
+--jit.off()
+
 function love.conf(t)
 	t.window.title = "quadSlice tests"
+	t.window.resizable = true
 end